+++ ddl/postgres/i2b2demodata/provider_dimension.sql
@@ -10,7 +10,7 @@ CREATE TABLE provider_dimension (
     download_date timestamp without time zone,
     import_date timestamp without time zone,
     sourcesystem_cd character varying(50),
-    upload_id numeric(38,0)
+    upload_id integer
 );
 
 --
