+++ ddl/postgres/i2b2demodata/qt_patient_enc_collection.sql
@@ -12,11 +12,11 @@ CREATE SEQUENCE qt_sq_qper_pecid
 -- Name: qt_patient_enc_collection; Type: TABLE; Schema: i2b2demodata; Owner: -
 --
 CREATE TABLE qt_patient_enc_collection (
-    patient_enc_coll_id numeric(10,0) DEFAULT nextval('qt_sq_qper_pecid'::regclass) NOT NULL,
-    result_instance_id numeric(5,0),
-    set_index numeric(10,0),
-    patient_num numeric(10,0),
-    encounter_num numeric(10,0)
+    patient_enc_coll_id integer DEFAULT nextval('qt_sq_qper_pecid'::regclass) NOT NULL,
+    result_instance_id integer,
+    set_index integer,
+    patient_num integer,
+    encounter_num integer
 );
 
 --
