+++ ddl/postgres/i2b2demodata/qt_privilege.sql
@@ -5,6 +5,6 @@ CREATE TABLE qt_privilege (
     protection_label_cd character varying(1500),
     dataprot_cd character varying(1000),
     hivemgmt_cd character varying(1000),
-    plugin_id numeric(10,0)
+    plugin_id integer
 );
